requirements.txt
functions-framework==3.*
xarray
numpy
pandas
pvlib
cartopy
matplotlib
pymongo
google.cloud